@@ -26,20 +26,21 @@
         }
     ],
     "require": {
-        "php": "^7.2",
+        "php": "^7.4 || ^8.0",
         "ext-json": "*",
-        "league/uri-interfaces": "^2.1",
+        "league/uri-interfaces": "^2.3",
         "psr/http-message": "^1.0"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.3",
-        "guzzlehttp/psr7": "^1.4",
+        "friendsofphp/php-cs-fixer": "^v3.8.0",
+        "guzzlehttp/psr7": "^2.2",
         "league/uri": "^6.0",
-        "phpstan/phpstan": "^0.12",
-        "phpstan/phpstan-phpunit": "^0.12",
-        "phpstan/phpstan-strict-rules": "^0.12",
-        "phpunit/phpunit": "^7.0 | ^8.0",
-        "laminas/laminas-diactoros": "^2.0"
+        "phpstan/phpstan": "^1.7",
+        "phpstan/phpstan-phpunit": "^1.1",
+        "phpstan/phpstan-strict-rules": "^1.2",
+        "phpstan/phpstan-deprecation-rules": "^1.0",
+        "phpunit/phpunit": "^9.5.20",
+        "laminas/laminas-diactoros": "^2.11.0"
     },
     "suggest": {
         "ext-intl": "to handle IDN host",
@@ -62,25 +63,19 @@
     },
     "scripts": {
         "phpcs": "php-cs-fixer fix -v --diff --dry-run --allow-risky=yes --ansi",
-        "phpstan-src": "phpstan analyse -l max -c phpstan.src.neon src --ansi",
-        "phpstan-tests": "phpstan analyse -l max -c phpstan.tests.neon tests --ansi",
-        "phpstan": [
-            "@phpstan-src",
-            "@phpstan-tests"
-        ],
+        "phpcs:fix": "php-cs-fixer fix -vvv --allow-risky=yes --ansi",
+        "phpstan": "phpstan analyse -l max -c phpstan.neon src --memory-limit=256M --ansi",
         "phpunit": "phpunit --coverage-text",
         "test": [
-            "@phpcs",
-            "@phpstan-src",
-            "@phpstan-tests",
-            "@phpunit"
+            "@phpunit",
+            "@phpstan",
+            "@phpcs:fix"
         ]
     },
     "scripts-descriptions": {
         "phpcs": "Runs coding style test suite",
+        "phpcs:fix": "Fix coding style",
         "phpstan": "Runs complete codebase static analysis",
-        "phpstan-src": "Runs source code static analysis",
-        "phpstan-test": "Runs test suite static analysis",
         "phpunit": "Runs unit and functional testing",
         "test": "Runs full test suite"
     },
