@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Validation\Constraint;
 
@@ -9,15 +10,14 @@ use Lcobucci\JWT\Validation\ConstraintViolation;
 final class IssuedBy implements Constraint
 {
     /** @var string[] */
-    private $issuers;
+    private array $issuers;
 
-    /** @param list<string> $issuers */
-    public function __construct(...$issuers)
+    public function __construct(string ...$issuers)
     {
         $this->issuers = $issuers;
     }
 
-    public function assert(Token $token)
+    public function assert(Token $token): void
     {
         if (! $token->hasBeenIssuedBy(...$this->issuers)) {
             throw new ConstraintViolation(
