@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Validation\Constraint;
 
@@ -8,15 +9,14 @@ use Lcobucci\JWT\Validation\ConstraintViolation;
 
 final class PermittedFor implements Constraint
 {
-    /** @var string  */
-    private $audience;
+    private string $audience;
 
-    public function __construct($audience)
+    public function __construct(string $audience)
     {
         $this->audience = $audience;
     }
 
-    public function assert(Token $token)
+    public function assert(Token $token): void
     {
         if (! $token->isPermittedFor($this->audience)) {
             throw new ConstraintViolation(
