@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Validation\Constraint;
 
@@ -8,15 +9,14 @@ use Lcobucci\JWT\Validation\ConstraintViolation;
 
 final class RelatedTo implements Constraint
 {
-    /** @var string */
-    private $subject;
+    private string $subject;
 
-    public function __construct($subject)
+    public function __construct(string $subject)
     {
         $this->subject = $subject;
     }
 
-    public function assert(Token $token)
+    public function assert(Token $token): void
     {
         if (! $token->isRelatedTo($this->subject)) {
             throw new ConstraintViolation(
