@@ -1,13 +1,11 @@
 <?php
 
 /**
- * League.Uri (http://uri.thephpleague.com/components)
+ * League.Uri (https://uri.thephpleague.com/components/2.0/)
  *
  * @package    League\Uri
  * @subpackage League\Uri\Components
  * @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
- * @license    https://github.com/thephpleague/uri-components/blob/master/LICENSE (MIT License)
- * @version    2.0.2
  * @link       https://github.com/thephpleague/uri-components
  *
  * For the full copyright and license information, please view the LICENSE
@@ -37,21 +35,14 @@ final class UserInfo extends Component implements UserInfoInterface
 
     private const REGEXP_ENCODED_CHAR = ',%[A-Fa-f0-9]{2},';
 
-    /**
-     * @var string|null
-     */
-    private $user;
-
-    /**
-     * @var string|null
-     */
-    private $pass;
+    private ?string $user;
+    private ?string $pass;
 
     /**
      * New instance.
      *
-     * @param mixed|null $user
-     * @param mixed|null $pass
+     * @param object|float|int|string|bool|null $user
+     * @param object|float|int|string|bool|null $pass
      */
     public function __construct($user = null, $pass = null)
     {
