@@ -5,7 +5,7 @@
         'type' => 'library',
         'install_path' => __DIR__ . '/../',
         'aliases' => array(),
-        'reference' => '9971eda8cc6d5a3e2675c4a2876a391453c5b492',
+        'reference' => 'c93060337623a22b30e5e03a7e6b10e629530aea',
         'name' => '__root__',
         'dev' => false,
     ),
@@ -16,7 +16,7 @@
             'type' => 'library',
             'install_path' => __DIR__ . '/../',
             'aliases' => array(),
-            'reference' => '9971eda8cc6d5a3e2675c4a2876a391453c5b492',
+            'reference' => 'c93060337623a22b30e5e03a7e6b10e629530aea',
             'dev_requirement' => false,
         ),
         'beberlei/assert' => array(
@@ -262,13 +262,22 @@
             'reference' => 'ad87d5a5ca981228e0e205c2bc7dfb8e24559b60',
             'dev_requirement' => false,
         ),
+        'lcobucci/clock' => array(
+            'pretty_version' => '2.0.0',
+            'version' => '2.0.0.0',
+            'type' => 'library',
+            'install_path' => __DIR__ . '/../lcobucci/clock',
+            'aliases' => array(),
+            'reference' => '353d83fe2e6ae95745b16b3d911813df6a05bfb3',
+            'dev_requirement' => false,
+        ),
         'lcobucci/jwt' => array(
-            'pretty_version' => '3.4.6',
-            'version' => '3.4.6.0',
+            'pretty_version' => '4.1.5',
+            'version' => '4.1.5.0',
             'type' => 'library',
             'install_path' => __DIR__ . '/../lcobucci/jwt',
             'aliases' => array(),
-            'reference' => '3ef8657a78278dfeae7707d51747251db4176240',
+            'reference' => 'fe2d89f2eaa7087af4aa166c6f480ef04e000582',
             'dev_requirement' => false,
         ),
         'league/event' => array(
@@ -308,12 +317,12 @@
             'dev_requirement' => false,
         ),
         'league/uri-components' => array(
-            'pretty_version' => '2.2.1',
-            'version' => '2.2.1.0',
+            'pretty_version' => '2.4.1',
+            'version' => '2.4.1.0',
             'type' => 'library',
             'install_path' => __DIR__ . '/../league/uri-components',
             'aliases' => array(),
-            'reference' => '14bab20b05bc20d16dec17401a9027c679412589',
+            'reference' => '06f49e726a7fd06063e80f5f8b045fb9128e6c43',
             'dev_requirement' => false,
         ),
         'league/uri-interfaces' => array(
