@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Encoding;
 
@@ -8,12 +9,7 @@ use RuntimeException;
 
 final class CannotEncodeContent extends RuntimeException implements Exception
 {
-    /**
-     * @param JsonException $previous
-     *
-     * @return self
-     */
-    public static function jsonIssues(JsonException $previous)
+    public static function jsonIssues(JsonException $previous): self
     {
         return new self('Error while encoding to JSON', 0, $previous);
     }
