@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Signer;
 
@@ -7,12 +8,7 @@ use Lcobucci\JWT\Exception;
 
 final class CannotSignPayload extends InvalidArgumentException implements Exception
 {
-    /**
-     * @pararm string $error
-     *
-     * @return self
-     */
-    public static function errorHappened($error)
+    public static function errorHappened(string $error): self
     {
         return new self('There was an error while creating the signature: ' . $error);
     }
