@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Signer;
 
@@ -7,18 +8,12 @@ use Lcobucci\JWT\Exception;
 
 final class InvalidKeyProvided extends InvalidArgumentException implements Exception
 {
-    /**
-     * @param string $details
-     *
-     * @return self
-     */
-    public static function cannotBeParsed($details)
+    public static function cannotBeParsed(string $details): self
     {
         return new self('It was not possible to parse your key, reason: ' . $details);
     }
 
-    /** @return self */
-    public static function incompatibleKey()
+    public static function incompatibleKey(): self
     {
         return new self('This key is not compatible with this signer');
     }
