@@ -1,52 +1,69 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Signer\Key;
 
-use Lcobucci\JWT\Encoding\CannotDecodeContent;
 use Lcobucci\JWT\Signer\Key;
+use Lcobucci\JWT\SodiumBase64Polyfill;
+use SplFileObject;
+use Throwable;
 
-use function base64_decode;
+use function assert;
+use function is_string;
 
-final class InMemory extends Key
+final class InMemory implements Key
 {
-    /**
-     * @param string $contents
-     * @param string $passphrase
-     *
-     * @return self
-     */
-    public static function plainText($contents, $passphrase = '')
+    private string $contents;
+    private string $passphrase;
+
+    private function __construct(string $contents, string $passphrase)
+    {
+        $this->contents   = $contents;
+        $this->passphrase = $passphrase;
+    }
+
+    public static function empty(): self
+    {
+        return new self('', '');
+    }
+
+    public static function plainText(string $contents, string $passphrase = ''): self
     {
         return new self($contents, $passphrase);
     }
 
-    /**
-     * @param string $contents
-     * @param string $passphrase
-     *
-     * @return self
-     */
-    public static function base64Encoded($contents, $passphrase = '')
+    public static function base64Encoded(string $contents, string $passphrase = ''): self
     {
-        $decoded = base64_decode($contents, true);
-
-        if ($decoded === false) {
-            throw CannotDecodeContent::invalidBase64String();
-        }
+        $decoded = SodiumBase64Polyfill::base642bin(
+            $contents,
+            SodiumBase64Polyfill::SODIUM_BASE64_VARIANT_ORIGINAL
+        );
 
         return new self($decoded, $passphrase);
     }
 
-    /**
-     * @param string $path
-     * @param string $passphrase
-     *
-     * @return InMemory
-     *
-     * @throws FileCouldNotBeRead
-     */
-    public static function file($path, $passphrase = '')
+    /** @throws FileCouldNotBeRead */
+    public static function file(string $path, string $passphrase = ''): self
     {
-        return new self('file://' . $path, $passphrase);
+        try {
+            $file = new SplFileObject($path);
+        } catch (Throwable $exception) {
+            throw FileCouldNotBeRead::onPath($path, $exception);
+        }
+
+        $contents = $file->fread($file->getSize());
+        assert(is_string($contents));
+
+        return new self($contents, $passphrase);
+    }
+
+    public function contents(): string
+    {
+        return $this->contents;
+    }
+
+    public function passphrase(): string
+    {
+        return $this->passphrase;
     }
 }
