@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Signer\Key;
 
@@ -8,25 +9,46 @@ use function file_exists;
 use function strpos;
 use function substr;
 
-/** @deprecated Use \Lcobucci\JWT\Signer\Key\InMemory::file() instead */
-final class LocalFileReference extends Key
+/** @deprecated please use {@see InMemory} instead */
+final class LocalFileReference implements Key
 {
-    const PATH_PREFIX = 'file://';
+    private const PATH_PREFIX = 'file://';
 
-    /**
-     * @param string $path
-     * @param string $passphrase
-     *
-     * @return self
-     *
-     * @throws FileCouldNotBeRead
-     */
-    public static function file($path, $passphrase = '')
+    private string $path;
+    private string $passphrase;
+    private string $contents;
+
+    private function __construct(string $path, string $passphrase)
+    {
+        $this->path       = $path;
+        $this->passphrase = $passphrase;
+    }
+
+    /** @throws FileCouldNotBeRead */
+    public static function file(string $path, string $passphrase = ''): self
     {
         if (strpos($path, self::PATH_PREFIX) === 0) {
             $path = substr($path, 7);
         }
 
-        return new self(self::PATH_PREFIX . $path, $passphrase);
+        if (! file_exists($path)) {
+            throw FileCouldNotBeRead::onPath($path);
+        }
+
+        return new self($path, $passphrase);
+    }
+
+    public function contents(): string
+    {
+        if (! isset($this->contents)) {
+            $this->contents = InMemory::file($this->path)->contents();
+        }
+
+        return $this->contents;
+    }
+
+    public function passphrase(): string
+    {
+        return $this->passphrase;
     }
 }
