@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Token;
 
@@ -7,28 +8,17 @@ use Lcobucci\JWT\Exception;
 
 final class InvalidTokenStructure extends InvalidArgumentException implements Exception
 {
-    /** @return self */
-    public static function missingOrNotEnoughSeparators()
+    public static function missingOrNotEnoughSeparators(): self
     {
         return new self('The JWT string must have two dots');
     }
 
-    /**
-     * @param string $part
-     *
-     * @return self
-     */
-    public static function arrayExpected($part)
+    public static function arrayExpected(string $part): self
     {
         return new self($part . ' must be an array');
     }
 
-    /**
-     * @param string $value
-     *
-     * @return self
-     */
-    public static function dateIsNotParseable($value)
+    public static function dateIsNotParseable(string $value): self
     {
         return new self('Value is not in the allowed date format: ' . $value);
     }
