@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Token;
 
@@ -9,15 +10,10 @@ use function sprintf;
 
 final class RegisteredClaimGiven extends InvalidArgumentException implements Exception
 {
-    const DEFAULT_MESSAGE = 'Builder#withClaim() is meant to be used for non-registered claims, '
+    private const DEFAULT_MESSAGE = 'Builder#withClaim() is meant to be used for non-registered claims, '
                                   . 'check the documentation on how to set claim "%s"';
 
-    /**
-     * @param string $name
-     *
-     * @return self
-     */
-    public static function forClaim($name)
+    public static function forClaim(string $name): self
     {
         return new self(sprintf(self::DEFAULT_MESSAGE, $name));
     }
