@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Token;
 
@@ -7,8 +8,7 @@ use Lcobucci\JWT\Exception;
 
 final class UnsupportedHeaderFound extends InvalidArgumentException implements Exception
 {
-    /** @return self */
-    public static function encryption()
+    public static function encryption(): self
     {
         return new self('Encryption is not supported yet');
     }
