@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Validation;
 
@@ -7,5 +8,5 @@ use Lcobucci\JWT\Token;
 interface Constraint
 {
     /** @throws ConstraintViolation */
-    public function assert(Token $token);
+    public function assert(Token $token): void;
 }
