@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Validation\Constraint;
 
@@ -7,8 +8,7 @@ use Lcobucci\JWT\Exception;
 
 final class LeewayCannotBeNegative extends InvalidArgumentException implements Exception
 {
-    /** @return self */
-    public static function create()
+    public static function create(): self
     {
         return new self('Leeway cannot be negative');
     }
