@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Lcobucci\JWT\Validation;
 
@@ -6,7 +7,7 @@ use Lcobucci\JWT\Token;
 
 final class Validator implements \Lcobucci\JWT\Validator
 {
-    public function assert(Token $token, Constraint ...$constraints)
+    public function assert(Token $token, Constraint ...$constraints): void
     {
         if ($constraints === []) {
             throw new NoConstraintsGiven('No constraint given.');
@@ -28,7 +29,7 @@ final class Validator implements \Lcobucci\JWT\Validator
         Constraint $constraint,
         Token $token,
         array &$violations
-    ) {
+    ): void {
         try {
             $constraint->assert($token);
         } catch (ConstraintViolation $e) {
@@ -36,7 +37,7 @@ final class Validator implements \Lcobucci\JWT\Validator
         }
     }
 
-    public function validate(Token $token, Constraint ...$constraints)
+    public function validate(Token $token, Constraint ...$constraints): bool
     {
         if ($constraints === []) {
             throw new NoConstraintsGiven('No constraint given.');
