@@ -1,13 +1,11 @@
 <?php
 
 /**
- * League.Uri (http://uri.thephpleague.com/components)
+ * League.Uri (https://uri.thephpleague.com/components/2.0/)
  *
  * @package    League\Uri
  * @subpackage League\Uri\Components
  * @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
- * @license    https://github.com/thephpleague/uri-components/blob/master/LICENSE (MIT License)
- * @version    2.0.2
  * @link       https://github.com/thephpleague/uri-components
  *
  * For the full copyright and license information, please view the LICENSE
@@ -27,7 +25,11 @@ use function array_pop;
 use function array_reduce;
 use function end;
 use function explode;
+use function gettype;
 use function implode;
+use function is_object;
+use function is_string;
+use function method_exists;
 use function sprintf;
 use function strpos;
 use function substr;
@@ -43,10 +45,17 @@ final class Path extends Component implements PathInterface
         %(?![A-Fa-f0-9]{2}))
     /x';
 
+    private string $path;
+
     /**
-     * @var string
+     * New instance.
+     *
+     * @param object|float|int|string|bool|null $path
      */
-    private $path;
+    public function __construct($path = '')
+    {
+        $this->path = $this->validate($path);
+    }
 
     /**
      * {@inheritDoc}
@@ -57,19 +66,9 @@ final class Path extends Component implements PathInterface
     }
 
     /**
-     * New instance.
-     *
-     * @param mixed|string $path
-     */
-    public function __construct($path = '')
-    {
-        $this->path = $this->validate($path);
-    }
-
-    /**
      * Validate the component content.
      *
-     * @param mixed $path the path content
+     * @param object|float|int|string|bool|null $path the path content
      *
      * @throws TypeError if the component is no valid
      */
@@ -81,6 +80,24 @@ final class Path extends Component implements PathInterface
         }
 
         throw new TypeError('The path can not be null.');
+    }
+
+    /**
+     * Returns a new instance from an string or a stringable object.
+     *
+     * @param string|object $path
+     */
+    public static function createFromString($path = ''): self
+    {
+        if (is_object($path) && method_exists($path, '__toString')) {
+            return new self((string) $path);
+        }
+
+        if (!is_string($path)) {
+            throw new TypeError(sprintf('The path must be a string or a stringable object value, `%s` given', gettype($path)));
+        }
+
+        return new self($path);
     }
 
     /**
@@ -115,6 +132,14 @@ final class Path extends Component implements PathInterface
     public function getContent(): ?string
     {
         return $this->encodeComponent($this->path, self::REGEXP_PATH_ENCODING);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getUriComponent(): string
+    {
+        return (string) $this->getContent();
     }
 
     /**
